// Front-End/movies-streaming-app/src/app/core/interfaces/IPersonDetails.ts
export interface IPersonDetailsList {
    page : number,
    results : IPersonDetails[],
    total_pages: number,
    total_results: number
}

export interface IPersonDetails {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    known_for: KnownFor[];
}

export interface KnownFor {
    backdrop_path: string;
    id: number;
    title: string;
    original_title: string;
    overview: string;
    poster_path: string;
    media_type: string;
    adult: boolean;
    original_language: string;
    genre_ids: number[];
    popularity: number;
    release_date: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}