// Front-End/movies-streaming-app/src/app/core/interfaces/IMedia.ts
export interface IMediaList {
    page: number;
    results: IMedia[];
    total_pages: number;
    total_results: number;
}
export interface IMedia {
    backdrop_path: string | null;
    id: number;
    title: string;
    original_title: string;
    overview: string;
    poster_path: string;
    media_type: string;
    adult: boolean;
    original_language: string;
    genre_ids: number[];
    popularity: number;
    release_date: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
}